updates + fix tests
