+++ packages/fuel-indexer-tests/Cargo.toml
@@ -38,6 +38,7 @@ fuels = { version = "0.30", features = ["fuel-core-lib"] }
 fuels-abigen-macro = "0.30"
 fuels-core = "0.30"
 futures = "0.3"
+hex = "0.4"
 itertools = "0.10"
 rand = "0.8"
 reqwest = { version = "0.11", default-features = false, features = ["json", "rustls-tls"] }
