+++ packages/fuel-indexer-tests/trybuild/fail_if_attribute_abi_arg_includes_invalid_type.stderr
@@ -32,7 +32,7 @@ error[E0412]: cannot find type `BlockData` in this scope
   |
 help: consider importing one of these items
    |
-2  | use fuel_indexer_plugin::types::fuel::BlockData;
+2  | use fuel_indexer_plugin::prelude::fuel::BlockData;
    |
 2  | use fuel_indexer_types::abi::BlockData;
    |
@@ -45,7 +45,7 @@ error[E0422]: cannot find struct, variant or union type `BlockData` in this scop
   |
 help: consider importing one of these items
    |
-2  | use fuel_indexer_plugin::types::fuel::BlockData;
+2  | use fuel_indexer_plugin::prelude::fuel::BlockData;
    |
 2  | use fuel_indexer_types::abi::BlockData;
    |
@@ -58,7 +58,7 @@ error[E0422]: cannot find struct, variant or union type `TransactionData` in thi
   |
 help: consider importing one of these items
    |
-2  | use fuel_indexer_plugin::types::fuel::TransactionData;
+2  | use fuel_indexer_plugin::prelude::fuel::TransactionData;
    |
 2  | use fuel_indexer_types::abi::TransactionData;
    |
@@ -71,7 +71,7 @@ error[E0433]: failed to resolve: use of undeclared type `TransactionStatus`
   |
 help: consider importing one of these items
    |
-2  | use fuel_indexer_plugin::types::tx::TransactionStatus;
+2  | use fuel_indexer_plugin::prelude::tx::TransactionStatus;
    |
 2  | use fuel_indexer_types::tx::TransactionStatus;
    |
@@ -89,7 +89,7 @@ error[E0433]: failed to resolve: use of undeclared type `Transaction`
   |
 help: consider importing one of these items
    |
-2  | use fuel_indexer_plugin::types::tx::Transaction;
+2  | use fuel_indexer_plugin::prelude::tx::Transaction;
    |
 2  | use fuel_indexer_types::tx::Transaction;
    |
